+++ docs/example/Input/demos/formTest.tsx
@@ -7,6 +7,8 @@ import { Button, Input } from '@banana/banana-react';
 import React from 'react';
 
 export default function FormTest() {
+  const [controlledInputValue, setControlledInputValue] = React.useState('controlled');
+
   return (
     <div>
       <form
@@ -22,11 +24,30 @@ export default function FormTest() {
         <Input type="password" name="password" />
         <Input type="text" name="test" value="123" disabled />
         <Input type="email" name="email" required placeholder="required email" />
+        <Input
+          controlled
+          type="text"
+          name="controlled"
+          value={controlledInputValue}
+          placeholder="controlled input"
+          onChange={(e: any) => {
+            console.log('change');
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-unsafe-member-access
+            setControlledInputValue(e.detail.value);
+          }}
+        />
         <Button name="submit" value="haha" htmlType="submit">
           submit
         </Button>
         <Button htmlType="reset">reset</Button>
         <button type="submit">native submit</button>
+        <Button
+          onClick={() => {
+            setControlledInputValue('123');
+          }}
+        >
+          change the value of controlled input to '123'
+        </Button>
       </form>
     </div>
   );